consul-notifier/requirements.txt
python-docker-py
